requirements.txt
modal
streamlit
requests
Pillow
torch
numpy
fastapi
uvicorn
python-multipart
